Allow granular selection of permissions
Previously the only action that could be turned off is enabling the
GCP APIs, and all of the other modifications were done all the time.
This commit adds several flags to control different modifications, and
adds an "--enable_all" flag that will allow users to keep the old
behavior. It also changes -e to --enable_all from --enable_apis.