Properly label namespaces in migration tests
For tests where we're migrating from OSS, we never re-label the
namespaces with the correct revision before doing a rolling restart.
With this PR, we put all of the script output info a FIFO and grep out
what the new label is supposed to be, and then label the namespace
properly before performing a restart of the workloads.