Add kubeconfig (GKE only) support to create-mesh subcommand.
The subcommand can accept N kubeconfig files, each of which is expected
to contain a default context that will be the ONLY context used. The UX
for N clusters would be to supply N differennt files, each of which will
have a default context set to the corresponding target cluster.