Check all possible managed revisions for the validation webhook (#995)
Instead of just checking for asm-managed, check for asm-managed-rapid,
asm-managed, and asm-managed-stable in that order, and use the first
webhook that exists as the validating webhook.

This will allow the managed control plane to return a subset of
revisions (for code that uses new features, e.g. fleet memberships),
instead of always needing to create the asm-managed revision.