Handle kubeconfig via Connect gateway
Customers can also generate kubeconfig files by using the Connect
gateway instead of the usual gcloud container command. In this case, we
can't parse any information out of it nor can we validate the GCP
settings (or even tell if it's GKE or not really).

If we see that the kubeconfig context name starts with "connectgateway",
we now assume that it's actually connecting via Connect gateway and
don't attempt to parse any info or validate GCP resources.